setup for static export to work with IPFS and fix webpack lib issues
