+++ app/components/index.ts
@@ -1,1 +1,2 @@
 export * from "./Swapbox";
+export * from "./SwapboxContainer";
